fastAPI-trigget-bat/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
azure-storage-blob==12.19.0
python-dotenv==1.0.0
